Remove old app menu in favor of new one
